Refactor AI to be enabled dynamically
