+++ bot/generic.py
@@ -1,0 +1,41 @@
+import logging
+import re
+
+import coloredlogs
+
+
+class LogFilter(logging.Filter):
+    def filter(self, record: logging.LogRecord):
+        regexs = [
+            r"^Shard ID (%s) has sent the (\w+) payload\.$",
+            r"^Got a request to (%s) the websocket\.$",
+            r"^Shard ID (%s) has connected to Gateway: (%s) \(Session ID: (%s)\)\.$",
+            r"^Shard ID (%s) has successfully (\w+) session (%s) under trace %s\.$",
+            r"^Websocket closed with (%s), attempting a reconnect\.$",
+        ]
+        # 0 means block, anything else (e.g. 1) means allow
+        return next((0 for regex in regexs if re.match(regex, str(record.msg))), 1)
+
+
+TEXT_FORMAT = "%(asctime)s %(name)s %(levelname)s %(message)s"
+handler = logging.FileHandler(filename="discord.log", encoding="utf-8", mode="w")
+handler.setFormatter(logging.Formatter(TEXT_FORMAT))
+
+logger = logging.getLogger("bert")
+logger.addHandler(handler)
+coloredlogs.install(
+    level=logging.DEBUG,
+    logger=logger,
+    fmt=TEXT_FORMAT,
+)
+
+pycord_logger = logging.getLogger("discord")
+pycord_logger.addHandler(handler)
+coloredlogs.install(
+    level=logging.INFO,
+    logger=pycord_logger,
+    fmt=TEXT_FORMAT,
+)
+
+for pycord_handler in pycord_logger.handlers:
+    pycord_handler.addFilter(LogFilter())
